Extract tcp helper
